Fix win32 build

+++ src/common/Utility.h
@@ -100,7 +100,7 @@ private:
 #endif // !_DEBUG
 
 constexpr std::size_t operator "" _z(unsigned long long n) {
-  return n;
+  return static_cast<std::size_t> (n);
 }
 
 constexpr std::ptrdiff_t operator "" _sz(unsigned long long n) {
